+++ src/frontend/src/common/constants.js
@@ -1,1 +1,3 @@
 export const INGREDIENT_COUNTER_LIMIT_MAX = 3;
+export const LOCAL_STORAGE_APP_STATE = "_1651675_app_state";
+export const LOCAL_STORAGE_SOURCE_DATA = "_1651675_source_data";
