+++ src/frontend/src/layouts/AppLayoutDefault.vue
@@ -1,0 +1,14 @@
+<template>
+  <section>
+    <header></header>
+    <slot></slot>
+  </section>
+</template>
+
+<script>
+export default {
+  name: "AppLayoutDefault",
+};
+</script>
+
+<style></style>
